chore(open-api): reduce diff
